chore: update exclusions logic
