+++ packages/amplify-migration-tests/src/__tests__/migration_tests_v10/auth-add-all.migration.test.ts
@@ -11,35 +11,9 @@ import {
     removeAuthWithDefault,
   } from '@aws-amplify/amplify-e2e-core';
 import { versionCheck, allowedVersionsToMigrateFrom } from '../../migration-helpers';
+import { cfnDiffExclusions } from '../../migration-helpers-v10/cfn-diff-exclusions';
 import { initJSProjectWithProfileV10 } from '../../migration-helpers-v10/init';
-import { assertNoParameterChangesBetweenProjects, collectCloudformationDiffBetweenProjects, ExcludeFromCFNDiff } from '../../migration-helpers/utils';
-
-/**
- * Due to a known limitation of APIGateway, we do not want the AWS::ApiGateway::Deployment resource
- * from being compared in the CFN diff because it is regenerated whenever an amplify app is pulled down.
- * This would produce a false positive when checking for differences in the CFN templates of project1 & project2.
- * https://github.com/aws/aws-cdk/issues/8646#issuecomment-647561856
- */
-const cfnDiffExclusions: ExcludeFromCFNDiff = (currentCategory: string, currentResourceKey: string, cfnTemplates: {
-  project1: any;
-  project2: any;
-}) => {
-  const excludeAPIGateWayDeploymentResource = (cfnTemplate: any): any => {
-    const resources = cfnTemplate.Resources ?? {};
-    const resourceKeys = Object.keys(resources);
-    for(let key of resourceKeys){
-      const resource = resources[key];
-      if(resource.Type === 'AWS::ApiGateway::Deployment'){
-        delete resources[key];
-      }
-    }
-  }
-  if(currentCategory === 'api'){
-    excludeAPIGateWayDeploymentResource(cfnTemplates.project1);
-    excludeAPIGateWayDeploymentResource(cfnTemplates.project2);
-  }
-  return { project1: cfnTemplates.project1, project2: cfnTemplates.project2 };
-}
+import { assertNoParameterChangesBetweenProjects, collectCloudformationDiffBetweenProjects } from '../../migration-helpers/utils';
   
 describe('amplify migration test auth', () => {
     let projRoot1: string;
